Move initial sensor code to own folder
