+++ lib/sensors_old/LSM6DSO32.cpp
@@ -1,4 +1,4 @@
-#include "LSM6DSO32.h"
+#include "../LSM6DSO32.h"
 
 LSM6DSO32Sensor::LSM6DSO32Sensor(String n) 
     : Sensor(n)
