+++ lib/sensors_old/MAX_M10S.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 #include <Wire.h>
 #include <SparkFun_u-blox_GNSS_Arduino_Library.h>
-#include "sensorClass.h"
+#include "../sensorClass.h"
 
 class MAXM10SSensor : public Sensor {
 public:
