+++ lib/sensors_old/kx134.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 #include <SPI.h>
-#include "sensorClass.h"  
+#include "../sensorClass.h"
 
 class KX134Sensor : public Sensor {
 public:
